Organized SendProcess into a class. Not sure if it compiles, but methods need implementing.
